backend/requirements.txt
# =============================================================================
# MeetMind Backend — Dependencies
# =============================================================================

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Datenbank
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.3

# Konfiguration
pydantic-settings==2.5.2

# Datei-Handling
python-multipart==0.0.9
filetype==1.2.0

# Auth (Clerk JWT-Validierung)
PyJWT==2.9.0
cryptography==43.0.1
httpx==0.27.2

# Monitoring
sentry-sdk[fastapi]==2.14.0

# ---- Milestone 2: Transkription (bei Bedarf auskommentieren) ----
# openai-whisper==20240930

# ---- Milestone 3: Embeddings (bei Bedarf auskommentieren) ----
# sentence-transformers==3.2.1
# torch==2.4.1  # CPU-only

# ---- Milestone 4: RAG (bei Bedarf auskommentieren) ----
# langchain==0.3.7
# langchain-community==0.3.7
# langchain-ollama==0.2.1
# ragas==0.2.6

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Code-Qualität
ruff==0.6.9
black==24.10.0
